# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplyPatchResult": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "patch": {
            "$ref": "#/components/schemas/PatchOut"
          }
        },
        "required": [
          "ok",
          "message",
          "patch"
        ],
        "title": "ApplyPatchResult",
        "type": "object"
      },
...
      "ExceptionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExceptionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "open_count": {
            "title": "Open Count",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "open_count",
          "items"
        ],
        "title": "ExceptionList",
        "type": "object"
      },
      "ExceptionOut": {
        "properties": {
          "automation_id": {
            "title": "Automation Id",
            "type": "string"
          },
          "automation_name": {
            "default": "",
            "title": "Automation Name",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "human_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Decision"
          },
          "human_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Note"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_features": {
            "additionalProperties": true,
            "title": "Input Features",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "signature_key": {
            "title": "Signature Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "automation_id",
          "reason",
          "input_features",
          "signature_key",
          "confidence",
          "status",
          "human_decision",
          "human_note",
          "created_at"
        ],
        "title": "ExceptionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatchList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatchOut"
            },
            "title": "Items",
            "type": "array"
          },
          "proposed_count": {
            "title": "Proposed Count",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "proposed_count",
          "items"
        ],
        "title": "PatchList",
        "type": "object"
      },
      "PatchOut": {
        "properties": {
          "auto_applicable": {
            "title": "Auto Applicable",
            "type": "boolean"
          },
          "automation_id": {
            "title": "Automation Id",
            "type": "string"
          },
          "automation_name": {
            "default": "",
            "title": "Automation Name",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "from_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Value"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "proposed_by": {
            "title": "Proposed By",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "rule": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "to_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Value"
          }
        },
        "required": [
          "id",
          "automation_id",
          "kind",
          "step_id",
          "field",
          "from_value",
          "to_value",
          "confidence",
          "auto_applicable",
          "status",
          "rationale",
          "rule",
          "evidence_count",
          "proposed_by",
          "created_at"
        ],
        "title": "PatchOut",
        "type": "object"
      },
...
      "ResolveExceptionRequest": {
        "properties": {
          "decision": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ResolveExceptionRequest",
        "type": "object"
      },
      "ResolveExceptionResult": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "rules_proposed": {
            "title": "Rules Proposed",
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "message",
          "rules_proposed"
        ],
        "title": "ResolveExceptionResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Detects repetitive enterprise workflows from activity logs, converts them into automations, and promotes those automations from suggested to autonomous through a measured trust ladder.",
    "title": "LOOP \u2014 Workflow Intelligence Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exceptions": {
      "get": {
        "description": "The human queue: executions the automation was not confident about.",
        "operationId": "list_exceptions_api_v1_exceptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|resolved)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Exceptions",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/exceptions/{exception_id}/resolve": {
      "post": {
        "description": "Record a human decision, then look for a rule worth learning.\n\nThe decision is the training signal: enough matching decisions on the same\ninput shape and LOOP proposes the branch that would have handled them.",
        "operationId": "resolve_exception_api_v1_exceptions__exception_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "exception_id",
            "required": true,
            "schema": {
              "title": "Exception Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveExceptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveExceptionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Exception",
        "tags": [
          "governance"
        ]
      }
    },
...
    "/api/v1/patches": {
      "get": {
        "description": "Self-healing proposals and learned rules awaiting review.",
        "operationId": "list_patches_api_v1_patches_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(proposed|applied|rejected)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Patches",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/patches/{patch_id}/apply": {
      "post": {
        "description": "Apply a proposed patch to the flow definition.",
        "operationId": "apply_patch_api_v1_patches__patch_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "patch_id",
            "required": true,
            "schema": {
              "title": "Patch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyPatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Patch",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/patches/{patch_id}/reject": {
      "post": {
        "description": "Dismiss a proposal without changing the flow.",
        "operationId": "reject_patch_api_v1_patches__patch_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "patch_id",
            "required": true,
            "schema": {
              "title": "Patch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyPatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Patch",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}